feat(category): add validation to prevent deletion of categories with related products
- Add check in delete method to verify if category has related products
- Throw ValidationError if category cannot be deleted

+++ src/services/category.service.ts
@@ -1,13 +1,17 @@
 import { NotFoundError } from '../errors/not-found.erro.js';
+import { ValidationError } from '../errors/validation.error.js';
 import { Category } from '../models/category.model.js';
 import { CategoryRepository } from '../repositories/category.repository.js';
+import { ProductRepository } from "../repositories/product.repository.js";
 
 export class CategoryService {
 
   private categoryRepository: CategoryRepository;
+  private productRepository: ProductRepository;
 
   constructor() {
     this.categoryRepository = new CategoryRepository();
+    this.productRepository = new ProductRepository();
   }
 
   async getAll(): Promise<Category[]> {
@@ -35,7 +39,10 @@ export class CategoryService {
     await this.categoryRepository.update(_category);
   }
 
-  async delete(id: string): Promise<void> {
+  async delete(id: string) {
+    if (await this.productRepository.getCountByCategoria(id) > 0) {
+        throw new ValidationError("Não é possível excluir uma categoria com produtos relacionados!");
+    }
     await this.categoryRepository.delete(id);
-  }
+}
 }
